requirements.txt
flask==2.3.2
pydantic==2.1.1
PyPDF2==3.0.1
redis[hiredis]==4.6.0
scholarly==1.7.11
lingua-language-detector==1.3.2
pymongo[srv]==4.4.1
python-decouple==3.8